+++ src/Stormancer.Plugins/Friends/Stormancer.Server.Plugins.Friends/Changelog.rst
@@ -9,6 +9,14 @@ Please use only reStructuredText in this file, no Markdown!
 
 This project adheres to semantic versioning.
 
+Unreleased
+----------
+Changed
+*******
+- Updated Users dependency to 5.x
+- Added a CancellationToken parameter to most async APIs.
+- Use new S2S protocols internally and use code generator to create S2S proxy.
+
 2.0.2
 -----
 Changed