+++ src/Stormancer.Plugins/GameFinder/Stormancer.Server.Plugins.GameFinder/Changelog.rst
@@ -8,6 +8,16 @@ The format is based on `Keep a Changelog <https://keepachangelog.com/en/1.0.0/>`
 Please use only reStructuredText in this file, no Markdown!
 
 This project adheres to semantic versioning.
+
+Unreleased
+----------
+Added
+*****
+- Added GameFinderProxy.OpenGameSession method to tell a gamefinder that a gamesession needs more players.
+Changed
+*******
+- Update Users dependency to 5.x.
+
 6.0.0.3
 -------
 Changed