+++ src/Stormancer.Plugins/Steam/Stormancer.Server.Plugins.Steam/Changelog.rst
@@ -9,6 +9,12 @@ Please use only reStructuredText in this file, no Markdown!
 
 This project adheres to semantic versioning.
 
+Unreleased
+----------
+Changed
+*******
+- Update to Users 5.x
+
 4.0.3
 -----
 Changed